Added SquareDivisors

+++ src/SquareDivisors.java
@@ -2,8 +2,8 @@ public class SquareDivisors{
     static int squareFactor(long N){
         // code here
         int count = 0;
-        for(int i = 1; i <= Math.sqrt(N); i++){
-            if(N % (i * i) == 0){
+        for(long i = 1; i <= Math.sqrt(N); i++){
+            if(N % (i * i) == 0L){
                 count++;
             }
         }
